Implementation of Text Object & Bugfixes
The text object was fully implemented

An example program (rot13.asm) was added to the examples
directory. It uses TextWindow and Text objects and
primitive commands of the VM to perform ROT13
"encryption".

A bug in the assembler was fixed that didn't increment the
current address correctly when defining character sequences
with the command "char"

A bug in the vm was fixed that lead to grt, lrt, leq, geq,
equ and neq not behaving according to their specification

A fix was added to the Write function in the TextWindow
object so the text that is written will appear on
the display unconditionally.